third_party/blink/web_tests/platform/mac-mac10.10/external/wpt/html/rendering/widgets/button-layout/grid-expected.txt

This is a testharness.js-based test.
FAIL inline-grid assert_equals: clientWidth expected 26 but got 20
FAIL grid assert_equals: clientWidth expected 26 but got 20
Harness: the test ran to completion.
